% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_palettes.R
\docType{data}
\name{.required_strat_categories}
\alias{.required_strat_categories}
\title{Required stratigraphic categories}
\format{
An object of class \code{character} of length 30.
}
\usage{
.required_strat_categories
}
\description{
These are the canonical categories that all stratigraphic palettes
must cover. They are used by \code{\link[=validate_stratpal]{validate_stratpal()}} to ensure
palettes are complete.
}
\keyword{internal}
