% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_palettes.R
\docType{data}
\name{.stratpalettes}
\alias{.stratpalettes}
\title{Internal list of stratigraphic palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
.stratpalettes
}
\description{
This list collects all palettes so they can be looked up by name.
Add new palettes here as you create them.
}
\keyword{internal}
