% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_palettes.R
\docType{data}
\name{stratpalettes}
\alias{stratpalettes}
\alias{stratpal_rpg}
\alias{stratpal_grays}
\title{Stratigraphic palettes}
\format{
Named character vectors

An object of class \code{character} of length 30.

An object of class \code{character} of length 30.
}
\usage{
stratpal_rpg

stratpal_grays
}
\description{
Named character vectors of hex colors for stratigraphic plotting.
These palettes can be passed to \code{\link[=scale_fill_stratpal]{scale_fill_stratpal()}}.
}
\examples{
stratpal_rpg["volcanic"]
stratpal_grays["soil"]

}
\keyword{datasets}
