% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_avstrat}
\alias{theme_avstrat}
\title{Custom theme for stratigraphic plots}
\usage{
theme_avstrat(base_size = 11, base_family = "arial")
}
\arguments{
\item{base_size}{Base font size. Defaults to 11.}

\item{base_family}{Base font famil. Defaults to Arial.}
}
\value{
A \code{\link[ggplot2:theme]{ggplot2::theme()}} object.
}
\description{
A ggplot2 theme designed to work well with typical plot ouput from avstrat,
and matching the author's preferred aesthetics.
}
\examples{
# Apply a custom theme to one plot
ggstrat(df = example_data_strat, section_name = "21LSHD02") +
  theme_avstrat()

# Set the custom theme as default for all plots
ggplot2::theme_set(theme_avstrat())
ggstrat(df = example_data_strat, section_name = "21LSHD02")
}
